Add systemtap assisted tracing support
I wrote another vscode extension
https://github.com/FlorentRevest/vscode-systemtap-assistant this one
facilitates tracing kernel code, especially in the context of syzkaller
issues debugging.

Systemtap is required on the host to transpile .stp scripts into kernel
modules, and on the guest to load these kernel modules and set some
global variables such as the traced program's PID or the output network
interface name.

On the host side, we introduce a new task to "deploy" the systemtap
script, this is run by the systemtap extension every time a probe point
is added. This also needs to run after kernel builds to make sure the
kernel module loads against the new kernel. If needs be we can evolve
this logic later to also dynamically reload tracers on the target using
SSH.

On the guest side, this modifies autostart.sh such that if a tracer is
detected, it will wrap the autostart execution with it. That is, only
logs from children of autostart are logged by default, but this can be
trivially changed.

A systemtap syntax highlighter extension is also provided to simplify
modifying the generated scripts.

The rest of the heavy lifting is done in the extension, namely:
- adding context menus to add probe points:
  https://github.com/FlorentRevest/vscode-systemtap-assistant/blob/main/src/Tracer.ts
- providing logs received from the systemtap scrtip as a text document
  https://github.com/FlorentRevest/vscode-systemtap-assistant/blob/main/src/Logs.ts