Disable debugger spam if "errors are detected"
If a user opens a file that is not indexed by clangd, it will probably
fail to parse with a bunch of errors. These errors are collected under
the "Problems" tab. If after this the user tries to start a debugger
session with F5, VSCode shows a popup asking for confirmation that
debugging should be started even though "problems have been detected".
I've observed multiple users run away screaming "it's broken". It turns
out that this popup can safely be ignored with a simple setting. Use it
by default, this will make people's lives easier.