Add a gdb symbol index ELF section to vmlinux
When gdb wants to access a somewhat complex variable, it needs to parse
the entire DWARF and build a symbol index. This can be very slow (in the
order of 6 seconds or so). When using VSCode's integrated debugger setup
this is always the case for a first breakpoint (there's always a local
variable that the UI tries to access which is painfully slow) so that
makes the whole setup quite annoying to use... I found that this
operation could be cached once and for all using the gdb-add-index
command, which actually only takes a couple of seconds. It parses the
DWARF once and stuff the gdb index into an ELF section. Subsequent runs
of gdb are much faster and the interactive debugging experience is a lot
nicer. This trades ~6seconds on every first breakpoint against ~2 extra
seconds at the end of every build.