tasks: accept symbolic links for local*.sh
In some cases, it might be useful to use symbolic links for the
local*.sh files, e.g. to use one from another repository where the file
can be tracked and even shared.

Switching from -type to -xtype will do the trick: -xtype checks the type
of the file that -type does not check.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>