Fix SILENT_BUILD_FLAG in tasks.sh
When using := to assign to a variable in Bash, it tests for a parameter
that is unset or null. The empty string counts as null, so if I set
SILENT_BUILD_FLAG="" to make the build verbose this assignment will
still override it and make the build silent because Bash treats an emtpy
string as "null".