Replace mmdebstrap with mkosi to improve compatibility
Replace mmdebstrap with mkosi for rootfs generation to enhance portability.
This patch updates the tasks.sh script to use mkosi instead of mmdebstrap,
enabling support when mmdebstrap is not available for example on RHEL-based
distros. The mkosi command is configured to match mmdebstrap's functionality,
with an additional step to move the rootfs from the mkosi-generated 'image'
subdirectory to the expected directory, ensuring compatibility with the
script's structure.

Fixes #16

Signed-off-by: Ali Nasrolahi <A.Nasrolahi01@gmail.com>